Produce a Sankey-like graph from Insight `topn` data using Graphviz dot
This is an example of using the Insight API to extract data from Insight.

This program runs a query against the `topn` endpoint to get
data that can be used to build the relationship between Source IP address,
Source Port, Destination Port, and Destination IP address with traffic flow
rates represented on each link.

The two non-standard Python packages required are:
  - requests; http://docs.python-requests.org/
  - networkx: https://networkx.github.io/

And you will need the `dot` package from Graphviz (http://graphviz.org/) to
convert the output of this script into a graphical representation of the
traffic