This is an example of filtering mitigations by alert ID which applies to SP/SL 9.0 and prior. (#17)
An example of filtering mitigations by alert id

This was directly taken from a request from ATAC, and since I wrote the script, I figured it shouldn't go to waste.  So here it is.  It defaults to looking at one page of mitigations, so specify more pages on the command line if you want to see more mitigations.